rename and refactor servlets for home page, changed front css file for home
